refactored InputManager class from Menu class
